brain-tumor-project/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
sqlalchemy==2.0.31
alembic==1.13.1
pydantic==2.8.2
pydantic-settings==2.3.4
pillow==10.4.0
torch==2.3.1
torchvision==0.18.1
numpy==1.26.4
opencv-python-headless==4.10.0.84
scikit-learn==1.5.1
matplotlib==3.9.1
reportlab==4.2.2
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.27.0
